Aula_Curso_Guanabara

+++ Exercício_aula_final/ex006.py
@@ -1,4 +1,4 @@
-n1=int(input('Digite um número:'))
+n1=float(input('Digite um número:'))
 n2=n1*2
 n3=n1*3
 n4=n1**(1/2)
